Move some backend functions to traversal
